Update the Docker business logic to use the latest Docker API and polyglot.
Now, the Docker image that submissions are run in is a pre-built version of the
Dockerfile called 'ulysse/polyglot'.

The latest version of the Docker API no longer returns just a string when we
attach to the container. It now returns an array containing two arrays, the
first being an array of chunks sent to STDOUT, the second being an array of
chunks sent to STDERR. This commmit simply treats the output as STDERR
concatenated to STDOUT, which would allow teachers to specify rules for standard
out and not have to worry about buffer ordering crap.