Add Respond.js for IE compatability and change links to work with HTTPS.
It turns out that BS3 does not support IE8 (the most up-to-date version of
internet explorer that can run on Windows XP) right out of the box because it
(BS3) relies on media queries that IE8 does not support. Respond.js fixes this.

Also, the links were changed from being of the form

  http://www.foo.bar

To

  //www.foo.bar

Which means that the protocol the browser uses to get those resources will
change whether the site is visited from HTTP or HTTPS.