Override basic JOptionPane IO so that they can be tested in submissions.
This involves creating a new JOptionPane class that overrides #showMessageDialog
and #showInputDialog to use System.out and System.in, respectively. In addition,
the Docker image has to be updated to use this new class (which *must* be
precompiled before being uploaded to the Docker image), and the cmd_for method
must now add a parameter so that the `java` command goes as:

  java -Xbootclasspath/p:java-overrides FooBar

This java-overrides directory is where javax/swing/JOptionPane.class resides.