Add support for Java packages.
When a Java program is uploaded, the file is searched for a line beginning with
"package". If such a line is found, then the next token is used as a package
name passed to the `java` command.

In order for packages to work, a new directory called `submission_bin` (not
simply called `bin` because there already exists a `/bin` directory) is created
and `javac` is told to write into this new directory. If we don't specify this
output directory with the `-d` flag, then `javac` will not automatically create
directories.