Only allow resubmission if the number of attempts hasn't been exceeded.
This is implemented at the Pundit-level, so as long as views and controllers use
Pundit, the user will not know the difference.

This cannot really be implemented at the model level because a teacher could
always edit the max attempts to be less than submission's current num attempts,
which would thereby make the submission fail validations.