Only teachers can see how many submissions an assignment has.
This is implemented by checking the policy for the right to list submissions.

This bug did not appear for listing how many submissions there are for a
particular course, because the ungraded submission count is only shown for
courses the user is teaching, which necessarily implies the user is the teacher
of that course (and so can see its submissions).