Make list-users-table work for IE8.
Whereas Chrome seems to be able to treat a column width as a min-width, IE does
not unless I say to use a fixed table-style. But when I do that with the current
width (32px) it really makes the column 32px wide and the picture spills over.
So I adjusted the value to be 49px instead.

Another issue is that IE8 does not support the :first-of-type selector. Thus, I
had to use two nested :first-child selectors to find the first column of a
table. But hey! It works.