Add a basic enroll-in-course via secret key functionality.
Courses now have an enroll_key property that is unique to every course.
Enrolling into a course now requires using the enroll_courses page, which
provides a way of finding and enrolling into a course by secret key.

Additional tests on the policy for courses are necessary in order to make sure
students can't double-enroll or anything like that.