Use Rails-default IDs in assignment forms.
These were at one point changed so that Javascript used to display when the
grace period would end could identify certain DOM elements. The problem with
this approach is that Rails uses the IDs of form elements to create labels and
fill in data and whatnot.

Instead, the Javascript now uses the (consistent) IDs Rails generates in order
to do work on DOM elements without getting in the way of the framework.

With this commit, tests in assignment_pages_spec that were previously failing
due to unfindable due_time and grace_period fields will now pass.