Use spaces rather than tabs when displaying submissions' code.
This commit merely replaces tabs with four spaces in order to get a consistent
look for submissions over all browsers. In addition, this prevents code from
reaching all the way across the screen.

Four spaces was not a number chosen arbitarily -- it is the length of a tab on
many systems by default. Thus, submissions' code will now more closely resemble
what they looked like on a students' computer.