Enable ssh agent forwarding in the Vagrant VM.
This allows me to do Capistrano deploys from within Vagrant. This is important
because Capistrano itself has dependencies, and these should be specified in the
Gemfile. But I can't use the Gemfile unless I'm on a machine that is going to
run the Rails app, and the Vagrant VM is such a machine.