Move deploy tasks into custom_tasks.cap and used shared folders.
It turns out Capistrano was completely ignoring my tasks because I had them in
the wrong file. For this I blame Capistrano's issue with documenting whether a
feature is for Capistrano 2 or Capistrano 3. In this case, I was using v2 syntax
when in fact I'm running v3.

In the future, all tasks will need to be placed in lib/capistrano/tasks/*.cap.

In addition, the public/system directory is explicitly symlinked. This will make
user submissions persist across all deployments.