Move Submission#execute_code! into a DJ worker.
This involves moving the many pieces of code dedicated to working on a
submission (located mostly in Submission and SubmissionsHelper) into the
SubmissionExecutionJob class.

With this new change, the execute_code! method now defers to a background job.
However, this means that specs must now reload submissions after calling
to-date with the changes the worker is making.

The long-term goal of this change is to make the worker more manageable. In this
commit a new log specifically for DJ workers is made (log/dj.log), and any
errors that occur there will be easier to debug.