Refactor SubmissionExecutionJob.
Two improvements are made. The first is that only @submission_id is stored
permanently, making serialization simpler and easier to debug in the future.

The second is that the job now logs when it begins, ends, gets a status report,
and when it has any errors. This too makes debugging easier.