Only create visible HTML for activities that'll be shown.
Public activity in this application is a bit of a hack, and logic for
determining whether or not an activity should be shown is being implemented at
the view level, which is bad. A future refactoring will allow this commit to be
reverted, but for now this will hide the underlying dirtiness of my
implementation.