Validate for valid UTF-8 in source files.
A common mistake for submitters will be to attempt to upload files that are not
valid UTF-8. This will break my views (who will do things like replacing tabs
with fixed numbers of spaces and other string operations on source code), and is
always a mistake on the part of the user.

If a source file is not valid UTF-8, then it's probably an already-compiled
class file or executable. These are obviously not what we want to work with.