Make first and last name work from the views.
This commit makes an important shift in how users are created. Devise now takes
care of creating users (it was already managing sessions before this commit).
There is a lot of useless code that will never be called still in the codebase,
but the refactoring necessary to clean this up is the subject of a different
commit or pull request.