Fix a bug in submission encoding validation.
An issue with the previous way I was doing it is that it won't quite work when
the file has not yet been placed in its final directory. This might happen if
the submission in question has not yet been saved, and thus the /000/000/001/
part of the path (the id here would be `1`) would not be a valid directory.

This new implementation favors the queued_for_write[:original] path, which is in
a tmp directory, over the final directory. It seems that which directory is
chosen depends on whether the submission exists yet or not.